add items feature in attack mode, fix get_tower bug

+++ assets/Scripts/Enemy_Info.cs
@@ -3,11 +3,11 @@ using System.Collections;
 
 public class Enemy_Info {
 	public string model;
-	public int health;
+	public float health;
 	public float speed;
 	public int number;
 	public int money;
-	public Enemy_Info(string model,int health,float speed,int number,int money){
+	public Enemy_Info(string model,float health,float speed,int number,int money){
 		this.model = model;
 		this.health = health;
 		this.speed = speed;
